refactor/mongo config 파일 config.mongo에 정리
